requirements.txt
streamlit
python-dotenv
langchain
openai
faiss-cpu
unstructured[local-inference]
pillow
PyPDF2
tiktoken
langchain-community